update features value check and test for it.
`features` was originally defaulted to `None`, which then we tested 
against to ensure users passed the correct value.  When we updated the 
default `features` to use all features, we did not update the value 
check for it in _init_ (`if features is None:`). Now we can switch that 
to `if not features` to ensure that some features were passed. Other 
tests had to be edited to not pass an empty list as the features.