Remove TensorFlow as a dependency. (#9)
* update features value check and test for it.

`features` was originally defaulted to `None`, which then we tested  against to ensure users passed the correct value.  When we updated the default `features` to use all features, we did not update the value check for it in _init_ (`if features is None:`). Now we can switch that to `if not features` to ensure that some features were passed. Other tests had to be edited to not pass an empty list as the features.

* remove keras.backend.image_data_format call from cell_tracker.

Change the default `data_format` from None (where we automatically get  the data format from keras) to "channels_last".  On _init_, check to make sure that data_format is a valid value.

* Replace tensorflow testing tools with pytest.

* Replace K.floatx() calls with new cell_tracker.dtype attribute (defaults to float32).

* Remove tensorflow from setup.py and requirements.txt.

* Added `absl-py` to travis builds and to setup.py extra_requires, previously installed by tensorflow.